requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
aiohttp>=3.9.0  # For async capabilities
ratelimit>=2.2.1  # For rate limiting 